Project_02_Smart_Doc_Uploader/backend/requirements.txt
# Project 02 Backend — Python Dependencies

# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pgvector>=0.3.0

# Pydantic + settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File handling
aiofiles>=23.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0

# Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
